moving all main logic into caddy package
